requirements.txt
python-dotenv~=1.1.1
Flask~=3.1.1
fastapi~=0.116.1
protobuf~=5.29.5